feat:完善User接口，设置Problem的身份验证

+++ src/main/kotlin/cn/kzoj/routes/ProblemRoutes.kt
@@ -3,20 +3,28 @@ package cn.kzoj.routes
 import cn.kzoj.data.problem.ProblemService
 import cn.kzoj.exception.problem.ProblemIdNotIntException
 import cn.kzoj.exception.problem.ProblemPageIndexNotPositiveIntException
+import cn.kzoj.exception.user.UserAuthorityException
 import cn.kzoj.models.problem.Problem
+import cn.kzoj.models.user.UserAuthority
+import cn.kzoj.models.user.UserSession
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.*
+import io.ktor.server.auth.authenticate
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
+import io.ktor.server.sessions.get
+import io.ktor.server.sessions.sessions
 
 fun Application.problemRoutes(problemService: ProblemService) {
     routing {
         route("/problem") {
             // CRUD
-            createProblem(problemService)
-            deleteProblem(problemService)
-            updateProblem(problemService)
+            authenticate("auth-session-user") {
+                createProblem(problemService)
+                deleteProblem(problemService)
+                updateProblem(problemService)
+            }
             queryProblemById(problemService)
             queryProblemByTitle(problemService)
             queryProblemByPage(problemService)
@@ -49,9 +57,19 @@ fun Route.createProblem(problemService: ProblemService) {
 fun Route.deleteProblem(problemService: ProblemService) {
     delete("/{id}") {
         val id = call.parameters["id"].toString().toIntOrNull()
-
         if (id == null) {
             throw ProblemIdNotIntException()
+        }
+
+        // 权限校验，仅管理员或创建者可操作
+        val userSession = call.sessions.get<UserSession>()
+        if (userSession == null) {
+            throw UserAuthorityException()
+        }
+        if (userSession.userAuthority != UserAuthority.ADMIN &&
+            problemService.queryProblemById(id).createdByUser != userSession.userId
+        ) {
+            throw UserAuthorityException()
         }
 
         problemService.deleteProblem(id)
@@ -65,9 +83,25 @@ fun Route.deleteProblem(problemService: ProblemService) {
  *
  * 接收题目: Problem
  */
+// TODO: Problem的更新也要通过Session实现控制，前端先访问query以显示题目编辑页面，此时后端设置Session，编辑完成后调用update，后端对Session进行校验，防止用户篡改题目id
 fun Route.updateProblem(problemService: ProblemService) {
     put("/update") {
         val newProblem = call.receive<Problem>()
+        // TODO: 可能存在篡改id破坏数据库的行为
+        if (newProblem.id == null) {
+            throw ProblemIdNotIntException()
+        }
+
+        // 权限校验，仅管理员或创建者可操作
+        val userSession = call.sessions.get<UserSession>()
+        if (userSession == null) {
+            throw UserAuthorityException()
+        }
+        if (userSession.userAuthority != UserAuthority.ADMIN &&
+            problemService.queryProblemById(newProblem.id).createdByUser != userSession.userId
+        ) {
+            throw UserAuthorityException()
+        }
 
         problemService.updateProblem(newProblem)
